add support for calculating "encapsulation"
A useful property to know for how much we should trust the volume
measurement is whether a cell is fully "encapsulated", i.e. whether
the locality around the cell fully surrounds that cell. This
happens whenever a cell is far enough from other cells, pillars,
and the edge of the FOV. This makes the volume measurement more
likely to be correct.